Remove long and double from type system

+++ modules/ballerina-core/src/main/java/org/ballerinalang/model/values/BInteger.java
@@ -27,20 +27,20 @@ import org.ballerinalang.model.types.BTypes;
  */
 public final class BInteger extends BValueType {
 
-    private int value;
+    private long value;
 
-    public BInteger(int value) {
+    public BInteger(long value) {
         this.value = value;
     }
 
     @Override
     public int intValue() {
-        return this.value;
+        return (int) this.value;
     }
 
     @Override
     public long longValue() {
-        return (long) this.value;
+        return this.value;
     }
 
     @Override
@@ -60,7 +60,7 @@ public final class BInteger extends BValueType {
 
     @Override
     public String stringValue() {
-        return Integer.toString(value);
+        return Long.toString(value);
     }
 
     @Override
